chore: improve NumberInput exports
